+++ cartoframes/utils/utils.py
@@ -161,6 +161,10 @@ def gen_variable_name(value):
     return 'v' + get_hash(value)[:6]
 
 
+def gen_column_name(value, operation=False):
+    return value if operation else '$' + value
+
+
 def get_hash(text):
     h = hashlib.sha1()
     h.update(text.encode('utf-8'))