+++ cartoframes/viz/helpers/color_category_layer.py
@@ -8,7 +8,7 @@ from ..layer import Layer
 def color_category_layer(
         source, value, title='', top=11, cat=None, palette=None,
         size=None, opacity=None, stroke_color=None, stroke_width=None,
-        description='', footer='', legend=True, popup=True,
+        description='', footer='', legend=True, popups=True,
         widget=False, animate=None, credentials=None):
     """Helper function for quickly creating a category color map.
 
@@ -32,8 +32,8 @@ def color_category_layer(
         footer (str, optional): Footer text placed under legend items.
         legend (bool, optional): Display map legend: "True" or "False".
           Set to "True" by default.
-        popup (bool, optional): Display popups on hover and click: "True" or "False".
-          Set to "True" by default.
+        popups (bool, list of :py:class:`Popup <cartoframes.viz.Popup>`, default False, optional):
+          Display popups on hover and click: "True" or "False". Set to "True" by default.
         widget (bool, optional): Display a widget for mapped data.
           Set to "False" by default.
         animate (str, optional): Animate features by date/time or other numeric field.
@@ -85,8 +85,8 @@ def color_category_layer(
                 'filter': animation_filter
             }
         },
-        popup=popup and not animate and get_popup(
-          popup, title, value, value),
+        popups=popups and not animate and get_popup(
+          popups, title, value, value),
         legend=legend and {
             'type': {
                 'point': 'color-category-point',