+++ cartoframes/viz/layer.py
@@ -6,6 +6,7 @@ from ..utils.utils import merge_dicts
 from .legend import Legend
 from .legend_list import LegendList
 from .popup import Popup
+from .popup_list import PopupList
 from .source import Source
 from .style import Style
 from .widget import Widget
@@ -14,7 +15,7 @@ from .widget_list import WidgetList
 from ..utils.utils import extract_viz_columns
 
 
-class Layer(object):
+class Layer():
     """Layer to display data on a map. This class can be used as one or more
     layers in :py:class:`Map <cartoframes.viz.Map>` or on its own in a Jupyter
     notebook to get a preview of a Layer.
@@ -25,10 +26,6 @@ class Layer(object):
             table name, SQL query or a dataframe.
         style (str, dict, or :py:class:`Style <cartoframes.viz.Style>`, optional):
             The style of the visualization.
-        popup (dict or :py:class:`Popup <cartoframes.viz.Popup>`, optional):
-            This option adds interactivity (click and hover) to a layer to show popups.
-            The columns to be shown must be added in a list format for each event.
-            See :py:class:`Popup <cartoframes.viz.Popup>` for more information.
         legends (:py:class:`Legend <cartoframes.viz.Legend>` list, optional):
             The legends definition for a layer. It contains a list of legend helpers.
             See :py:class:`Legend <cartoframes.viz.Legend>` for more information.
@@ -36,6 +33,10 @@ class Layer(object):
             Widget or list of widgets for a layer. It contains the information to display
             different widget types on the top right of the map. See
             :py:class:`WidgetList` for more information.
+        click_popup(`Popup <cartoframes.viz.Popup>`, optional): Set up a popup to be
+            displayed on a click event.
+        hover_popup(`Popup <cartoframes.viz.Popup>`, optional): Set up a popup to be
+            displayed on a hover event.
         credentials (:py:class:`Credentials <cartoframes.auth.Credentials>`, optional):
             A Credentials instance. This is only used for the simplified Source API.
             When a :py:class:`Source <cartoframes.viz.Source>` is passed as source,
@@ -53,81 +54,29 @@ class Layer(object):
         Create a layer with a custom popup, legend, and widget.
 
         .. code::
-
-            from cartoframes.auth import set_default_credentials
-            from cartoframes.viz import Layer
-
-            set_default_credentials(
-                base_url='https://cartovl.carto.com',
-                api_key='default_public'
-            )
-
-            Layer(
-                "SELECT * FROM populated_places WHERE adm0name = 'Spain'",
-                'color: ramp(globalQuantiles($pop_max, 5), reverse(purpor))',
-                popup={
-                    'hover': '$name',
-                    'click': ['$name', '$pop_max', '$pop_min']
-                },
-                legends={
-                    'type': 'color-category',
-                    'title': 'Population'
-                },
-                widgets=[{
-                    'type': 'formula',
-                    'title': 'Avg $pop_max',
-                    'value': 'viewportAvg($pop_max)'
-                }]
-            )
+            # FIXME
 
         Create a layer specifically tied to a :py:class:`Credentials
         <cartoframes.auth.Credentials>` and display it on a map.
 
         .. code::
-
-            from cartoframes.auth import Credentials
-            from cartoframes.viz import Layer, Map
-
-            credentials = Credentials(
-                base_url='https://cartovl.carto.com',
-                api_key='default_public'
-            )
-
-            pop_layer = Layer(
-                'populated_places',
-                'color: red',
-                credentials=credentials
-            )
-            Map(pop_layer)
+            # FIXME
 
         Preview a layer in a Jupyter notebook. Note: if in a Jupyter notebook,
         it is not required to explicitly add a Layer to a :py:class:`Map
         <cartoframes.viz.Map>` if only visualizing data as a single layer.
 
         .. code::
-
-            from cartoframes.auth import set_default_credentials
-            from cartoframes.viz import Layer, Map
-
-            set_default_credentials('https://cartoframes.carto.com')
-
-            pop_layer = Layer(
-                'brooklyn_poverty',
-                'color: ramp($poverty_per_pop, sunset)',
-                legends={
-                    'type': 'color-continuous',
-                    'title': 'Poverty per pop'
-                }
-            )
-            pop_layer
+            #FIXME
     """
 
     def __init__(self,
                  source,
                  style=None,
-                 popup=None,
                  legends=None,
                  widgets=None,
+                 click_popup=None,
+                 hover_popup=None,
                  credentials=None,
                  bounds=None,
                  geom_col=None):
@@ -136,14 +85,14 @@ class Layer(object):
 
         self.source = _set_source(source, credentials, geom_col)
         self.style = _set_style(style)
-        self.popup = _set_popup(popup)
+        self.popups = _set_popups({'click': click_popup, 'hover': hover_popup})
         self.legends = _set_legends(legends)
         self.widgets = _set_widgets(widgets)
 
         geom_type = self.source.get_geom_type()
-        popup_variables = self.popup.get_variables()
+        popups_variables = self.popups.get_variables()
         widget_variables = self.widgets.get_variables()
-        external_variables = merge_dicts(popup_variables, widget_variables)
+        external_variables = merge_dicts(popups_variables, widget_variables)
         self.viz = self.style.compute_viz(geom_type, external_variables)
         viz_columns = extract_viz_columns(self.viz)
 
@@ -152,7 +101,7 @@ class Layer(object):
         self.source_data = self.source.data
         self.bounds = bounds or self.source.bounds
         self.credentials = self.source.get_credentials()
-        self.interactivity = self.popup.get_interactivity()
+        self.interactivity = self.popups.get_interactivity()
         self.widgets_info = self.widgets.get_widgets_info()
         self.legends_info = self.legends.get_info() if self.legends is not None else None
         self.has_legend_list = isinstance(self.legends, LegendList)
@@ -182,14 +131,13 @@ def _set_style(style):
         return Style()
 
 
-def _set_popup(popup):
+def _set_popups(popups):
     """Set a Popup class from the input"""
-    if isinstance(popup, dict):
-        return Popup(popup)
-    elif isinstance(popup, Popup):
-        return popup
+
+    if isinstance(popups, (dict, Popup)):
+        return PopupList(popups)
     else:
-        return Popup()
+        return PopupList()
 
 
 def _set_legends(legends):