+++ cartoframes/viz/popups/__init__.py
@@ -1,0 +1,5 @@
+from .popup_element import popup_element
+
+__all__ = [
+  'popup_element'
+]