+++ tests/unit/viz/helpers/test_color_category_layer.py
@@ -31,12 +31,11 @@ class TestColorCategoryLayerHelper(object):
         assert layer.style._style['line']['color'] == 'opacity(ramp(top($name, 11), bold),1)'
         assert layer.style._style['polygon']['color'] == 'opacity(ramp(top($name, 11), bold), 0.9)'
 
-        assert layer.popup is not None
-        assert layer.popup._hover == [{
-            'title': 'Neighborhoods',
-            'value': '$name'
-        }]
+        assert layer.popups is not None
 
+        popup = layer.popups.elements[0]
+        assert popup.title == 'Neighborhoods'
+        assert popup.value == '$name'
         assert layer.legend is not None
         assert layer.legend._type['point'] == 'color-category-point'
         assert layer.legend._type['line'] == 'color-category-line'
@@ -145,21 +144,20 @@ class TestColorCategoryLayerHelper(object):
         layer = helpers.color_category_layer(
             self.source,
             'name',
-            popup=False
+            popups=False
         )
 
-        assert layer.popup._hover == []
+        assert len(layer.popups.elements) == 0
 
         layer = helpers.color_category_layer(
             self.source,
             'name',
-            popup=True
+            popups=True
         )
 
-        assert layer.popup._hover == [{
-            'title': 'name',
-            'value': '$name'
-        }]
+        popup = layer.popups.elements[0]
+        assert popup.title == 'name'
+        assert popup.value == '$name'
 
     def test_color_category_layer_widget(self, mocker):
         "should show/hide the widget"
@@ -190,7 +188,7 @@ class TestColorCategoryLayerHelper(object):
             animate='time'
         )
 
-        assert layer.popup._hover == []
+        assert len(layer.popups.elements) == 0
         assert layer.widgets._widgets[0]._type == 'time-series'
         assert layer.widgets._widgets[0]._title == 'Animation'
         assert layer.widgets._widgets[0]._value == 'time'