+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Migration Docs](/docs/developers/migrations/rc1.md)
 
 ### Changed
+- Remove `popup` parameter and add `hover_popup` and `click_popup` in Layer (#1348)
+- Add `popup_element` helpers (#1348)
 - Refactor Widgets API (#1349)
 - Refactor Legends API (#1347)